added base supervisor
