+++ test/test_macros.exs
@@ -1,7 +1,7 @@
 defmodule Cashier.TestMacros do
-    defmacro has_header(conn, header) do
-        quote do
-            Enum.member?(unquote(conn).req_headers, unquote(header))
-        end
+  defmacro has_header(conn, header) do
+    quote do
+      Enum.member?(unquote(conn).req_headers, unquote(header))
     end
+  end
 end
